requirements.txt
praw
python-dotenv
loguru
pyspark
kafka-python
requests
findspark